파트너 페이지가 뜰 수 있도록 처리함.
