+++ handler_partner.go
@@ -49,11 +49,29 @@ func handlePartners(w http.ResponseWriter, r *http.Request) {
 		Partnernum int       // 검색된 파트너사수
 		Devmode    bool      // 개발모드
 		SearchOption
-		Setting     Setting
-		Projectlist []string
+		Setting          Setting
+		Projectlist      []string
+		TasksettingNames []string
+		Stages           []Stage
+		Status           []Status
 	}
 	rcp := recipe{}
 	rcp.Projectlist, err = ProjectlistV2(client)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	rcp.TasksettingNames, err = TaskSettingNamesV2(client)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	rcp.Status, err = AllStatusV2(client)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	rcp.Stages, err = AllStagesV2(client)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
